requirements.txt
annotated-types==0.7.0
argcomplete==3.5.0
black==24.8.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
curl_cffi==0.7.1
datamodel-code-generator==0.25.9
dnspython==2.6.1
email_validator==2.2.0
genson==1.3.0
idna==3.7
inflect==5.6.2
isort==5.13.2
Jinja2==3.1.4
m3u8==6.0.0
MarkupSafe==2.1.5
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pyright==1.1.375
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
six==1.16.0
typed-argparse==0.3.1
typing_extensions==4.12.2
urllib3==2.2.2
